refactor: Move `KeyCode` to new module
