+++ src/event/key_code.rs
@@ -1,0 +1,249 @@
+use super::*;
+
+/// Represents a key.
+#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Clone, Copy, Hash)]
+#[cfg_attr(feature = "derive-more", derive(IsVariant))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub enum KeyCode {
+    /// Backspace key (Delete on macOS, Backspace on other platforms).
+    Backspace,
+    /// Enter key.
+    Enter,
+    /// Left arrow key.
+    Left,
+    /// Right arrow key.
+    Right,
+    /// Up arrow key.
+    Up,
+    /// Down arrow key.
+    Down,
+    /// Home key.
+    Home,
+    /// End key.
+    End,
+    /// Page up key.
+    PageUp,
+    /// Page down key.
+    PageDown,
+    /// Tab key.
+    Tab,
+    /// Shift + Tab key.
+    BackTab,
+    /// Delete key. (Fn+Delete on macOS, Delete on other platforms)
+    Delete,
+    /// Insert key.
+    Insert,
+    /// F key.
+    ///
+    /// `KeyCode::F(1)` represents F1 key, etc.
+    #[cfg_attr(feature = "derive-more", is_variant(ignore))]
+    F(u8),
+    /// A character.
+    ///
+    /// `KeyCode::Char('c')` represents `c` character, etc.
+    #[cfg_attr(feature = "derive-more", is_variant(ignore))]
+    Char(char),
+    /// Null.
+    Null,
+    /// Escape key.
+    Esc,
+    /// Caps Lock key.
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    CapsLock,
+    /// Scroll Lock key.
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    ScrollLock,
+    /// Num Lock key.
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    NumLock,
+    /// Print Screen key.
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    PrintScreen,
+    /// Pause key.
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    Pause,
+    /// Menu key.
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    Menu,
+    /// The "Begin" key (often mapped to the 5 key when Num Lock is turned on).
+    ///
+    /// **Note:** this key can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    KeypadBegin,
+    /// A media key.
+    ///
+    /// **Note:** these keys can only be read if
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] has been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    #[cfg_attr(feature = "derive-more", is_variant(ignore))]
+    Media(MediaKeyCode),
+    /// A modifier key.
+    ///
+    /// **Note:** these keys can only be read if **both**
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] and
+    /// [`KeyboardEnhancementFlags::REPORT_ALL_KEYS_AS_ESCAPE_CODES`] have been enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    #[cfg_attr(feature = "derive-more", is_variant(ignore))]
+    Modifier(ModifierKeyCode),
+}
+
+impl KeyCode {
+    /// Returns `true` if the key code is the given function key.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use crossterm::event::KeyCode;
+    /// assert!(KeyCode::F(1).is_function_key(1));
+    /// assert!(!KeyCode::F(1).is_function_key(2));
+    /// ```
+    pub fn is_function_key(&self, n: u8) -> bool {
+        matches!(self, KeyCode::F(m) if *m == n)
+    }
+
+    /// Returns `true` if the key code is the given character.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use crossterm::event::KeyCode;
+    /// assert!(KeyCode::Char('a').is_char('a'));
+    /// assert!(!KeyCode::Char('a').is_char('b'));
+    /// assert!(!KeyCode::F(1).is_char('a'));
+    /// ```
+    pub fn is_char(&self, c: char) -> bool {
+        matches!(self, KeyCode::Char(m) if *m == c)
+    }
+
+    /// Returns the character if the key code is a character key.
+    ///
+    /// Returns `None` if the key code is not a character key.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use crossterm::event::KeyCode;
+    /// assert_eq!(KeyCode::Char('a').as_char(), Some('a'));
+    /// assert_eq!(KeyCode::F(1).as_char(), None);
+    /// ```
+    pub fn as_char(&self) -> Option<char> {
+        match self {
+            KeyCode::Char(c) => Some(*c),
+            _ => None,
+        }
+    }
+
+    /// Returns `true` if the key code is the given media key.
+    ///
+    /// **Note:** this method requires
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] to be enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use crossterm::event::{KeyCode, MediaKeyCode};
+    /// assert!(KeyCode::Media(MediaKeyCode::Play).is_media_key(MediaKeyCode::Play));
+    /// assert!(!KeyCode::Media(MediaKeyCode::Play).is_media_key(MediaKeyCode::Pause));
+    /// ```
+    pub fn is_media_key(&self, media: MediaKeyCode) -> bool {
+        matches!(self, KeyCode::Media(m) if *m == media)
+    }
+
+    /// Returns `true` if the key code is the given modifier key.
+    ///
+    /// **Note:** this method requires both
+    /// [`KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES`] and
+    /// [`KeyboardEnhancementFlags::REPORT_ALL_KEYS_AS_ESCAPE_CODES`] to be enabled with
+    /// [`PushKeyboardEnhancementFlags`].
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use crossterm::event::{KeyCode, ModifierKeyCode};
+    /// assert!(KeyCode::Modifier(ModifierKeyCode::LeftShift).is_modifier(ModifierKeyCode::LeftShift));
+    /// assert!(!KeyCode::Modifier(ModifierKeyCode::LeftShift).is_modifier(ModifierKeyCode::RightShift));
+    /// ```
+    pub fn is_modifier(&self, modifier: ModifierKeyCode) -> bool {
+        matches!(self, KeyCode::Modifier(m) if *m == modifier)
+    }
+}
+
+impl Display for KeyCode {
+    /// Formats the `KeyCode` using the given formatter.
+    ///
+    /// # Platform-specific Notes
+    ///
+    /// On macOS, the Backspace key is displayed as "Delete", the Delete key is displayed as "Fwd
+    /// Del", and the Enter key is displayed as "Return". See
+    /// <https://support.apple.com/guide/applestyleguide/welcome/1.0/web>.
+    ///
+    /// On other platforms, the Backspace key is displayed as "Backspace", the Delete key is
+    /// displayed as "Del", and the Enter key is displayed as "Enter".
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            // On macOS, the Backspace key is called "Delete" and the Delete key is called "Fwd Del".
+            #[cfg(target_os = "macos")]
+            KeyCode::Backspace => write!(f, "Delete"),
+            #[cfg(target_os = "macos")]
+            KeyCode::Delete => write!(f, "Fwd Del"),
+
+            #[cfg(not(target_os = "macos"))]
+            KeyCode::Backspace => write!(f, "Backspace"),
+            #[cfg(not(target_os = "macos"))]
+            KeyCode::Delete => write!(f, "Del"),
+
+            #[cfg(target_os = "macos")]
+            KeyCode::Enter => write!(f, "Return"),
+            #[cfg(not(target_os = "macos"))]
+            KeyCode::Enter => write!(f, "Enter"),
+            KeyCode::Left => write!(f, "Left"),
+            KeyCode::Right => write!(f, "Right"),
+            KeyCode::Up => write!(f, "Up"),
+            KeyCode::Down => write!(f, "Down"),
+            KeyCode::Home => write!(f, "Home"),
+            KeyCode::End => write!(f, "End"),
+            KeyCode::PageUp => write!(f, "Page Up"),
+            KeyCode::PageDown => write!(f, "Page Down"),
+            KeyCode::Tab => write!(f, "Tab"),
+            KeyCode::BackTab => write!(f, "Back Tab"),
+            KeyCode::Insert => write!(f, "Insert"),
+            KeyCode::F(n) => write!(f, "F{n}"),
+            KeyCode::Char(c) => match c {
+                // special case for non-visible characters
+                ' ' => write!(f, "Space"),
+                c => write!(f, "{c}"),
+            },
+            KeyCode::Null => write!(f, "Null"),
+            KeyCode::Esc => write!(f, "Esc"),
+            KeyCode::CapsLock => write!(f, "Caps Lock"),
+            KeyCode::ScrollLock => write!(f, "Scroll Lock"),
+            KeyCode::NumLock => write!(f, "Num Lock"),
+            KeyCode::PrintScreen => write!(f, "Print Screen"),
+            KeyCode::Pause => write!(f, "Pause"),
+            KeyCode::Menu => write!(f, "Menu"),
+            KeyCode::KeypadBegin => write!(f, "Begin"),
+            KeyCode::Media(media) => write!(f, "{media}"),
+            KeyCode::Modifier(modifier) => write!(f, "{modifier}"),
+        }
+    }
+}
